Moved the DatabaseReader class from util.py to a separate file.
